Fixed Launcher addLink exception on CMS preview
If I opened a CMS page and clicked preview, I'd get the exception,
Invalid method Mage_Page_Block_Html_Breadcrumbs::addLink. This was
happening because the Launcher was trying to modify the layout of the
previewed page. This fix updates the Observer's _shouldBail method so
it skips pages that aren't in the adminhtml area.